% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{map_salmon1}
\alias{map_salmon1}
\title{map_salmon1}
\source{
Tsai 2016
}
\usage{
map_salmon1
}
\description{
Tsai 2016 (male recombination rates)
}
\author{
Torsten Pook \email{torsten.pook@uni-goettingen.de}
}
