% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{map_salmon2}
\alias{map_salmon2}
\title{map_salmon2}
\source{
Tsai 2016
}
\usage{
map_salmon2
}
\description{
Tsai 2016 (female recombination rates)
}
\author{
Torsten Pook \email{torsten.pook@uni-goettingen.de}
}
